Exec. lookup 'supported_roles' with wantlist=True
This way, we force the lookup function to return a list and then we
can check safely if a role is present in the lookup result.

Calling this lookup function without wantlist=True can lead to the
situation below if the function returns only one record:

+ 'setup_barman' in 'setup_barmanserver'

Which is true, and not the desired behavior.

With wantlist=True, the checking statement will be:

+ 'setup_barman' in ['setup_barmanserver']