Autotuning: configure vm.overcommit_ratio
With vm.swapiness=1 and vm.overcommit_memory=2, vm.overcommit_ratio
should allow the system to use the entire memory available.

                   RAM * vm.overcommit_ratio
Max Alloc = SWAP + -------------------------
                             100

If we want to allow the system to allocate the entire memory
(Max Alloc = RAM), then, overcommit_ratio is calculated as below:

                       RAM - SWAP
vm.overcommit_ratio = ------------ * 100
                          RAM

We ensure a minimum value of 50. This is usefull for the case when
the amount of swap memory is ridiculous higher than the RAM size.