Support Azure Database
Generalize the RDS specific paths as DBaaS since it can be shared with
Azure Database.

Azure Database has some unique usage with respect to how the username
needs to be generated in a connect strict as opposed to generate usage
in SQL statements.  HammerDB as well as some of the scripts needs to be
generated differently because of this.  There is currently an open
ticket with HammerDB to handle this.  This change addresses Azure
Database directly for ease of development.