dbt2: add tasks to update system packages
There is a library update that exists at some stage in the RHEL life
cycle that breaks some packages if they are not updated.  Unfortunately
package dependencies don't catch this and the most reliable way to not
run into this is to make sure the whole system is up to date.

One specific instance is with cmake.