Update system packages on RHEL8 based distros
There are a couple reasons we may need to force a system-wide package
update.

There is a bug in how one of packages is being built such that cmake
is broken.  Until the system image is updated, the simplest workaround
is to just update all of the installed packages.

Multiple repos are being used, and sometimes dependencies get confused
between packages from different repositories.  Making sure everything is
the from the latest view of each repository may be the easiest remedy.