Barman fixes
Do not execute the backup and check commands on standbys. Standby
nodes must be configured and ready to be backuped in case of
switchover/failover but are not supposed to be backuped when they
still are in recovery mode.

Use throttle=1 when updating the .pgpass file on the barman server
to avoid race condition that can lead to missing entries when working
with multiple PG nodes.