Execute yum updateinfo -y
Executing this command makes the import of GPG keys effective for
all the following dnf/yum executions.

If we don't do that, import of the keys will be done later by
another dnf/yum command. This shouldn't be an issue, but for unknown
reason, when dnf/yum imports keys then it does only that, whatever
the command being executed.

More details here:
https://bugzilla.redhat.com/show_bug.cgi?id=1768206