Cast use_hostname to bool
This is usually not necessary when use_hostname is defined in the
playbook file itself, or in an external json file. But, when
passing this boolean variable through the command line with:
-e 'use_hostname=false', in this case, the value is considered as
a string, then all the conditions based only on the value are true.

This fix may help in #306