Allow VIP usage with Barman
In the High-Availability context, when using a Virtual IP address for
routing read/write traffic to the primary node, then it's possible
to configure barman to use this VIP address as Postgres connection
endpoint. This way, in case of timeline change (following a
failover or switchover event), barman will be able to automatically
fetch new WAL records coming from the new primary node, without
requiring any manual operation.

This configuration is only possible when using the `postgres` backup
method, which is based on backup and WAL archiving using streaming
replication.

New hostvars:
- barman_private_vip: VIP address of the primary node. This VIP
  must be already active on the primary before configuring barman.
- barman_no_configuration: Disable the creation of barman
  configuration and backups for this instance. Must be set to
  `true` for all standby nodes in this mode.