Testing with multiple Ansible core versions (#436)
The test-runner.py script parses the ansible-core version, either passed in with flag --ansible-core-version or defined as an environment variable, ANSIBLE_CORE_VERSION, and then passes the parsed version variable to the testing functions. The version number has been added to the output results and logs.
In the Dockerfile.ansible-tester, the version of Debian has been upped to 11 to support ansible-core versions >= 2.12. The ARG ansible-core-version has been added to represent the variable being passed in at build time. This is defined in each docker-compose.yml to ensure each test case is being run with the correct ansible-core version.
In the Makefile, the flags --force-recreate --build has been added to ansible-tester-up to ensure that at each test case execution, a new ansible-tester build is being created. If this step is not included, the ansible-tester build uses the version of the first test case executed for every test case.
The default ansible-core version value has been set to 2.12.