Updated pgbackrest roles to allow more than 3 standby nodes.
Updated pgbackrest roles to allow more than 3 standby nodes. The tests for these roles was updated to ensure standby indexing was consistent and accurate.
An index for nodes was added to the /lookup/pgbackrest_nodes.py lookup function to have the indexing required in pgbackrest.conf stay consistent across all nodes. Using this index value, configuration lines are created and added to the end of the file.
On the standby nodes, this index value is accessible for pgbackrest.conf templating.