Module linesinfile
This in-house Ansible module is intended to provide better
performance than lineinfile when multiple lines must be managed.

Instead of using a loop, with linesinfile, we can pass an array
of lines. This way, the module is executed only once, whatever the
number of managed lines is.

Most of lineinfile's attributes are supported: line, regexp,
insertafter, insertbefore, state, backrefs, firstmatch, create,
owner, group, mode.

A few attributes are not supported: backup and search_string