Do not rewrite core_pattern
RHEL based distros come with tools to help list and capture core dumps
on the system and use systemd and journald facilities.
(systemd-coredump, coredumpctl, etc.)  If we rewrite core_pattern to
specifically not use systemd-coredump, then we cannot use these
facilities.

The only benefit to use not use them is if we want to rely on manually
handling all aspects of analyzing core dumps, or if we intend to
implement our own tools.