Remove kernel.core_pattern from sysctl.conf
Only if it's in there, which tpaexec will set.  tpaexec currently forces
some value to be set, and sometimes we want to use the distro's default
or another tool's preferred value that it may manage, like
systemd-coredump.  If we leave kernel.core_pattern in /etc/sysctl.conf,
tpaexec configures the system to run it last so it'll override anything
set before it.

This can be removed at a later time, when tpaexec can be used to not
set kernel.core_pattern.